requirements.txt
credentials
tweepy
selenium
Pillow